refactor: 계산기 로직 별도 ts 파일로 분리
